Move cors headers test to CorsV2IT
